refactor(clients): move Mounter, XOClient and NodeMetadataGetter clients into clients package
