+++ pkg/xenorchestra-csi/clients/node_metadata.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package xenorchestracsi
+package clients
 
 import (
 	"context"
